test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
flake8

# We need to pin Ansible to <2.4 until Zuul supports it, otherwise ansible-lint
# picks the latest version which is not known to work yet.
ansible<2.4
ansible-lint
bashate>=0.2
zuul-sphinx>=0.1.1
ruamel.yaml
